requirements.txt
streamlit
google-generativeai
python-dotenv
numpy
pandas
matplotlib
csvs-to-sqlite